AES CLI update
Modes of operation
Key input
