+++ include/aes_cli.h
@@ -1,5 +1,7 @@
+#ifndef _AES_CLI_H_
+#define _AES_CLI_H_
+
 #include <stdio.h>
-#include <string.h>
 #include <stdint.h>
 
 #define VERBOSE
@@ -17,47 +19,4 @@ typedef struct {
 } cli_ctx_t;
 
 void aes_cli (int argc, char** argv);
-void aes_help(void);
-void cli_mode (cli_ctx_t *cli_ctx, char *mode);
-void cli_action (cli_ctx_t *cli_ctx, char *act);
-void cli_file (cli_ctx_t *cli_ctx, char *path);
-void cli_key (cli_ctx_t *cli_ctx);
-void cli_aes_init (cli_ctx_t *cli_ctx);
-#include "../src/cli/aes_help.c"
-#include "../src/cli/cli_mode.c"
-#include "../src/cli/cli_action.c"
-#include "../src/cli/cli_file.c"
-#include "../src/cli/cli_key.c"
-#include "../src/cli/cli_aes_init.c"
-
-void aes_cli (int argc, char** argv) {
-    /* handle input
-    TODO add secription */
-
-    // CONTEXT INITIALIZATION
-    uint8_t key[CHUNK] = {0};
-    FILE *file_path = NULL;
-    cli_ctx_t cli_ctx = {ECB, ENCRYPT, key, file_path};
-
-    const char arg_help[] = "help";
-    if (argc == 2) {
-        if (strcmp(arg_help, argv[1]) == 0) aes_help();
-        else printf("Invalid input. Try typing \"aes help\"\n");
-        fclose(file_path);
-        return;
-    }
-
-    if (argc == 3) {
-        printf("Missing argument(s)\n");
-        fclose(file_path);
-        return;
-    }
-
-    cli_mode(&cli_ctx, argv[1]);     // MODE SELECTION
-    cli_action(&cli_ctx, argv[2]);   // ACTION
-    cli_file(&cli_ctx, argv[3]);     // FILE PATH
-    cli_key(&cli_ctx);               // KEY
-    cli_aes_init(&cli_ctx);          // ACTUAL ENCRYPTION/DECRYPTION
-
-    fclose(file_path);
-}
+#endif // !_AES_CLI_H_
