--- src/cli/aes_help.c
@@ -1,3 +1,0 @@
-void aes_help(void) {
-    // TODO: add help page
-}
